Add BudgetBar to CategoryDetail.kt
